requirements.txt
asgiref==3.8.1
Django==5.2.3
django-environ==0.12.0
sqlparse==0.5.3
